# Speed up test suite