Resolves https://github.com/stellar/stellar-core-internal/issues/371
Resolves https://github.com/stellar/stellar-core/issues/2100
Resolves #4445 

Looks like our test suite accumulated quite a lot of tech debt and is taking much longer than it should(this came up in a recent conversation with @drebelsky about simple tests like "txset nomination" taking ~35 seconds).

This PR makes several changes:

- Add a helper script to audit the test suite and identify biggest offenders (written by @claude)

The script can be run on all tests, any partition, or an individual test. sample output:

```
python3 run_test_timing_analysis.py -p overlay -n 10

Found 62 tests to run
Running each test individually to measure wall-clock time...
This may take a while depending on the number of tests...
--------------------------------------------------------------------------------
[1/62] Running: Flooding
[2/62] Running: ItemFetcher fetches
[3/62] Running: next peer strategy
...
[62/62] Running: Tracker works

================================================================================
TEST TIMING ANALYSIS REPORT (Wall-Clock Time)
================================================================================

Total tests: 62
Total time: 350.01s
Average time per test: 5.645s

Top 10 Slowest Tests:
--------------------------------------------------------------------------------
  1.  152.442s ( 43.6%) overlay flow control
  2.   50.374s ( 14.4%) TCPPeer read malformed messages
  3.   46.080s ( 13.2%) overlay pull mode loadgen
  4.   19.691s (  5.6%) survey request process order
  5.   14.766s (  4.2%) Flooding
  6.   11.077s (  3.2%) inbounds nodes can be promoted to ouboundvalid
  7.   10.106s (  2.9%) TCPPeer lifetime
  8.    9.474s (  2.7%) Time sliced static topology survey
  9.    6.130s (  1.8%) connecting to saturated nodes
 10.    5.087s (  1.5%) peer is purged from database after few failures

Top 10 tests account for 325.23s (92.9% of total time)

Warning: 1 tests failed
```

- One of the biggest issues is that we close a ledger every time we create a new account (some tests create thousands of accounts, which is slow even in virtual time). I switched those tests to use GENESIS_TEST_ACCOUNT_COUNT and observed some drastic speedups, e.g. "SCP Driver"

```
// Before
stellar-core test -a "SCP Driver"  30.92s user 49.76s system 37% cpu 3:35.51 total
// With GENESIS_TEST_ACCOUNT_COUNT
stellar-core test -a "SCP Driver"  1.89s user 0.20s system 73% cpu 2.840 total
```

- Move a couple of long-running tests to `acceptance`
- Tweak how load is generated in tests running in TCP mode, since it's running in real-time. Remaining TODO is to switch those to accelerated time. 
- Remove old upgrade tests - since the network is way past protocol 10, 11, and 12, replaying upgrades via catchup should be sufficient.


Opening as a draft, since I haven't finished auditing all the tests flagged by the scripts.