@marta-lokhova 2025-07-26
Here are top 50 slowest tests (a lot of them already live in `acceptance`, but some like `archival meta` run in every GH action). Note that these results are based on a regular core build; I imagine with asan - which is our default in CI - things are much worse.

```
Top 50 Slowest Tests:
--------------------------------------------------------------------------------
  1. 1392.434s ( 24.0%) History catchup
  2.  686.091s ( 11.8%) archival meta
  3.  482.937s (  8.3%) Catchup recent
  4.  417.904s (  7.2%) overlay parallel processing
  5.  287.317s (  5.0%) Publish catchup alternation with stall
  6.  264.985s (  4.6%) Catchup fatal failure
  7.  152.276s (  2.6%) overlay flow control
  8.  125.216s (  2.2%) LedgerTxnRoot prefetch classic entries
  9.  118.261s (  2.0%) ledger state update flow with parallel apply
 10.  117.431s (  2.0%) bucketmanager reattach to running merge
 11.  109.898s (  1.9%) write verified checkpoint hashes
 12.   80.054s (  1.4%) Retriggering catchups after trimming mSyncingLedgers
 13.   78.248s (  1.4%) History publish with restart
 14.   65.993s (  1.1%) Catchup failure recovery with buffered checkpoint
 15.   65.692s (  1.1%) catchup with a gap
 16.   58.802s (  1.0%) Publish throttles catchup
 17.   49.920s (  0.9%) TCPPeer read malformed messages
 18.   46.214s (  0.8%) overlay pull mode loadgen
 19.   41.728s (  0.7%) History bucket verification
 20.   39.476s (  0.7%) apply load
 21.   36.576s (  0.6%) generate load with unique accounts
 22.   34.024s (  0.6%) bl cache
 23.   32.347s (  0.6%) soroban cache population
 24.   31.407s (  0.5%) Catchup with protocol upgrade
 25.   29.497s (  0.5%) key-value lookup
 26.   29.390s (  0.5%) loadPoolShareTrustLinesByAccountAndAsset does not load outdated versions
 27.   29.347s (  0.5%) do not load outdated values
 28.   29.274s (  0.5%) bucket entry counters
 29.   29.259s (  0.5%) Flooding
 30.   29.251s (  0.5%) loadPoolShareTrustLinesByAccountAndAsset
 31.   27.614s (  0.5%) tx set hits overlay byte limit during construction
 32.   26.118s (  0.5%) surge pricing with DEX separation holds invariants
 33.   25.319s (  0.4%) BucketListIsConsistentWithDatabase bucket bounds
 34.   24.587s (  0.4%) create offer
 35.   23.430s (  0.4%) soroban txs accepted by the network
 36.   22.122s (  0.4%) Catchup manual
 37.   20.161s (  0.3%) parallel tx set building
 38.   20.112s (  0.3%) CKMSSample gamma distribution
 39.   19.430s (  0.3%) 3 nodes 2 running threshold 2
 40.   17.470s (  0.3%) pathpayment
 41.   16.893s (  0.3%) BucketListIsConsistentWithDatabase added entries
 42.   16.388s (  0.3%) survey request process order
 43.   15.033s (  0.3%) METADATA_DEBUG_LEDGERS works
 44.   15.033s (  0.3%) Soroban footprint validation
 45.   14.798s (  0.3%) getNodeWeight
 46.   14.457s (  0.2%) Soroban authorization
 47.   13.733s (  0.2%) History catchup with different modes
 48.   13.383s (  0.2%) txset nomination
 49.   11.820s (  0.2%) CKMSSample uniform distribution
 50.   11.771s (  0.2%) revoke from pool

Top 50 tests account for 5360.92s (92.5% of total time)
```