@marta-lokhova 2025-07-29
There's a surprising perf issue in "LedgerTxnRoot prefetch classic entries" test: [this](https://github.com/stellar/stellar-core/blob/62d0df174d50070bc3ff2a788f3157a4792723fe/src/ledger/test/LedgerTxnTests.cpp#L2742) call takes almost a minute, which seems way too high given we only generate 1000 valid classic entries.